Mandi Price/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
openai>=1.12.0
faiss-cpu>=1.7.4
pandas>=2.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
